% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{col2HSV}
\alias{col2HSV}
\title{col2HSV: converts a color to HSV in hexadecimal notation}
\usage{
col2HSV(color)
}
\arguments{
\item{color}{an R color name or a color in hexadecimal notation}
}
\value{
A character vector with the color(s) name(s) in hexadecimal notation
}
\description{
col2HSV converts an R color (or a set of colors) into an HSV color model, and
then returns the color names in hexadeciaml notation
}
\examples{

# convert 'tomato'
col2HSV("tomato")

}
\seealso{
\code{\link{setColors}}
}
\author{
Gaston Sanchez
}
