% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{.gridline_colour}
\alias{.gridline_colour}
\title{Derive a lighter gridline colour from text colour}
\usage{
.gridline_colour(textcol)
}
\description{
Derive a lighter gridline colour from text colour
}
\keyword{internal}
