% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{.palettes_cache}
\alias{.palettes_cache}
\title{Cached palette metadata}
\format{
An object of class \code{list} of length 23.
}
\usage{
.palettes_cache
}
\description{
Cached palette metadata
}
\keyword{internal}
