% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes_list}
\alias{palettes_list}
\title{Built-in palettes list}
\format{
An object of class \code{list} of length 23.
}
\usage{
palettes_list
}
\description{
Built-in palettes list
}
\keyword{internal}
