% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_any}
\alias{round_any}
\title{Round to nearest 'accuracy'}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{Numeric vector to round.}

\item{accuracy}{Granularity to round to.}

\item{f}{Rounding function (defaults to [base::round()]).}
}
\value{
A numeric vector with rounded values.
}
\description{
Round to nearest 'accuracy'
}
