% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_gradientn_manychat}
\alias{scale_color_gradientn_manychat}
\alias{scale_fill_gradientn_manychat}
\title{Manychat n-colour gradient scale}
\usage{
scale_color_gradientn_manychat(
  palette = "manychat_diverging_1",
  invert = FALSE,
  reverse = FALSE,
  ...
)

scale_fill_gradientn_manychat(
  palette = "manychat_diverging_1",
  invert = FALSE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Palette name; defaults to a Manychat palette.}

\item{invert}{Logical; invert colours via `1 - R/G/B`.}

\item{reverse}{Logical; reverse the colour order.}

\item{...}{Additional arguments passed to
\code{\link[ggplot2]{scale_color_gradientn}}.}
}
\description{
Continuous gradient scale using all colours from a palette, analogous to
\code{\link[ggplot2]{scale_color_gradientn}}.
}
