% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{show_palette}
\alias{show_palette}
\title{Show a palette in the viewer}
\usage{
show_palette(palette = NULL, reverse = FALSE, plot = TRUE)
}
\arguments{
\item{palette}{Character scalar naming one of the palettes defined in
`palettes.yml`.}

\item{reverse}{Logical; reverse the colour order.}

\item{plot}{Logical; draw palette?}
}
\description{
Show a palette in the viewer
}
