% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintalisation-package.R
\docType{package}
\name{vintalisation-package}
\alias{vintalisation}
\alias{vintalisation-package}
\title{vintalisation: Custom palettes and utilities for ggplot2}
\description{
This package provides branded colour palettes, ggplot2 scales/themes,
and helper utilities for generating consistent visualisations.
}
\keyword{internal}
