% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_plot}
\alias{make_plot}
\title{Save a ggplot into ./visualisations}
\usage{
make_plot(
  filename,
  plot = ggplot2::last_plot(),
  device = "png",
  height = 2500,
  scale = 16/9,
  unit = "px"
)
}
\arguments{
\item{filename}{Destination file name (without path).}

\item{plot}{Plot object to save.}

\item{device}{Graphics device to use (e.g. `"png"`).}

\item{height}{Height of the output file (default pixels).}

\item{scale}{Aspect ratio multiplier (defaults to 16/9).}

\item{unit}{Units for width/height, defaults to `"px"`.}
}
\description{
Save a ggplot into ./visualisations
}
