% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom minimal theme}
\usage{
theme_custom(
  market = "adevinta",
  base_size = 24,
  base_family = "DM Sans",
  heading_family = base_family,
  bg_fill = "#FFFFFF",
  blank_plot_bg = TRUE,
  legend_position = "bottom",
  show_y_axis = FALSE
)
}
\arguments{
\item{market}{One of "adevinta","marktplaats","gumtree","kijiji","2dehands"}

\item{base_size}{Base font size}

\item{base_family}{Base font family (defaults to "DM Sans")}

\item{heading_family}{Font family for titles/subtitles (defaults to
`base_family`).}

\item{bg_fill}{Plot background colour}

\item{blank_plot_bg}{Logical; if `TRUE`, draw the plot background using
`bg_fill`, otherwise keep it blank.}

\item{legend_position}{Position of the legend: "bottom", "top", "left",
"right", or "none".}

\item{show_y_axis}{Logical; if `TRUE`, display the Y-axis line.}
}
\description{
Custom minimal theme
}
