% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_manychat}
\alias{theme_manychat}
\title{Manychat branded theme}
\usage{
theme_manychat(
  market = "adevinta",
  base_size = 18/ggplot2::.pt,
  bg_fill = "#FFFFFF",
  blank_plot_bg = TRUE
)
}
\arguments{
\item{market}{One of "adevinta","marktplaats","gumtree","kijiji","2dehands"}

\item{base_size}{Base font size}

\item{bg_fill}{Plot background colour}

\item{blank_plot_bg}{Logical; if `TRUE`, draw the plot background using
`bg_fill`, otherwise keep it blank.}
}
\description{
Manychat branded theme
}
