% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{watermark}
\alias{watermark}
\title{Watermark layer}
\usage{
watermark(
  text = "Not validated",
  xmin = -Inf,
  xmax = Inf,
  ymin = -Inf,
  ymax = Inf,
  scale = 1,
  rot = 0,
  col = "grey",
  alpha = 0.5
)
}
\arguments{
\item{text}{Label to display as a watermark.}

\item{xmin, xmax, ymin, ymax}{Numeric bounds for the annotation. Defaults to
`-Inf`/`Inf` to span entire plot.}

\item{scale}{Scaling factor applied to the watermark text size.}

\item{rot}{Rotation angle in degrees.}

\item{col}{Watermark colour.}

\item{alpha}{Alpha transparency value.}
}
\value{
A ggplot2 annotation layer.
}
\description{
Adds a text watermark annotation to a ggplot. By default spans the entire
plot area.
}
